Add new BeautifulRobot operations and move Color and Pattern to BeautifulRobotDriver
