+++ src/main/java/frc/robot/subsystems/BeautifulRobot.java
@@ -7,11 +7,12 @@
 
 package frc.robot.subsystems;
 
-import edu.wpi.first.wpilibj.DriverStation;
 import edu.wpi.first.wpilibj.DriverStation.Alliance;
 import edu.wpi.first.wpilibj.command.Subsystem;
 import frc.robot.RobotMap;
 import frc.robot.misc.BeautifulRobotDriver.Operation;
+import frc.robot.misc.BeautifulRobotDriver.Pattern;
+import frc.robot.misc.BeautifulRobotDriver.Color;
 
 /**
  * BeautifulRobot&#8482; WS2812 RGB LED Strip Controller Subsystem.
@@ -85,31 +86,6 @@ public class BeautifulRobot extends Subsystem {
         setColor(Color.fromAlliance(color));
     }
 
-    /**
-     * A colour of the BeautifulRobot&#8482;.
-     */
-    public enum Color {
-        RED((byte) 0), BLUE((byte) 1), GREEN((byte) 2), YELLOW((byte) 3);
-
-        byte code;
-        Color(byte code) {
-            this.code = code;
-        }
-        public static Color fromAlliance(DriverStation.Alliance alliance) {
-            if(alliance == Alliance.Red) {
-                return Color.RED;
-            }
-            else if(alliance == Alliance.Blue) {
-                return Color.BLUE;
-            }
-            else {
-                return Color.GREEN;
-            }
-        }
-        public byte getCode() {
-            return code;
-        }
-    }
     private Color color = Color.GREEN;
 
     /**
@@ -150,20 +126,6 @@ public class BeautifulRobot extends Subsystem {
      */
     public boolean isOn() {
         return on;
-    }
-    /**
-     * All the different patterns the BeautifulRobot&#8482; has.
-     */
-    public enum Pattern {
-        SOLID((byte) 0), PULSATING((byte) 1), RAINBOW((byte) 2), MOVING_PULSE((byte) 3), PROGRESS_BAR((byte) 4);
-
-        private byte value;
-        Pattern(byte value) {
-            this.value = value;
-        }
-        public byte getValue() {
-            return value;
-        }
     }
 
     private Pattern pattern = Pattern.SOLID;
